+++ Modules/FarmLand/app/Http/Controllers/Api/V1/RehabilitationController.php
@@ -1,19 +1,21 @@
 <?php
 
-namespace Modules\FarmLand\Http\Controllers;
+namespace Modules\FarmLand\Http\Controllers\Api\V1;
+
 
 use App\http\Responses\ApiResponse;
 use App\Http\Controllers\Controller;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Modules\FarmLand\Http\Requests\Rehabilitations\StoreRehabilitationRequest;
 use Modules\FarmLand\Http\Requests\Rehabilitations\UpdateRehabilitationRequest;
 use Modules\FarmLand\Http\Resources\RehabilitationResource;
 use Modules\FarmLand\Models\Rehabilitation;
- 
+
 use Modules\FarmLand\Services\RehabilitationService;
 
 class RehabilitationController extends Controller
 {
-
+    use AuthorizesRequests;
     protected $rehabilitationService;
 
     public function __construct(RehabilitationService $rehabilitationService)
@@ -25,12 +27,13 @@ class RehabilitationController extends Controller
      */
     public function index()
     {
+        $this->authorize('viewAny', Rehabilitation::class);
         $data = $this->rehabilitationService->getAll();
-    return ApiResponse::success(
-        ['rehabilitations' => RehabilitationResource::collection($data)],
-        'Rehabilitation events retrieved successfully.',
-        200
-    );
+        return ApiResponse::success(
+            ['rehabilitations' => RehabilitationResource::collection($data)],
+            'Rehabilitation events retrieved successfully.',
+            200
+        );
     }
 
     /**
@@ -38,26 +41,30 @@ class RehabilitationController extends Controller
      */
     public function store(StoreRehabilitationRequest $request)
     {
+        $this->authorize('create', Rehabilitation::class);
         $rehabilitation = $this->rehabilitationService->store($request->validated());
 
-    return ApiResponse::success(
-        ['rehabilitations' => new RehabilitationResource($rehabilitation)],
-        'Rehabilitation events created successfully.',
-        201
-    );
+        return ApiResponse::success(
+            ['rehabilitations' => new RehabilitationResource($rehabilitation)],
+            'Rehabilitation events created successfully.',
+            201
+        );
     }
 
     /**
      * Display the specified resource.
      */
-    public function show( $id)
+    public function show($id)
     {
-          $rehabilitation = $this->rehabilitationService->getById($id);
-     return ApiResponse::success(
-[       ' rehabilitations' => $rehabilitation ,
-     ],  'Rehabilitation event retrieved successfully.',
-        200
-    );
+        $rehabilitation = $this->rehabilitationService->getById($id);
+        $this->authorize('view', $rehabilitation);
+        return ApiResponse::success(
+            [
+                ' rehabilitations' => $rehabilitation,
+            ],
+            'Rehabilitation event retrieved successfully.',
+            200
+        );
     }
 
     /**
@@ -65,13 +72,14 @@ class RehabilitationController extends Controller
      */
     public function update(UpdateRehabilitationRequest $request, Rehabilitation $rehabilitation)
     {
+        $this->authorize('update', $rehabilitation);
         $updated = $this->rehabilitationService->update($rehabilitation, $request->validated());
 
-     return ApiResponse::success(
-        ['rehabilitations' => new RehabilitationResource($updated)],
-        'Rehabilitation event updated successfully.',
-        200
-    );
+        return ApiResponse::success(
+            ['rehabilitations' => new RehabilitationResource($updated)],
+            'Rehabilitation event updated successfully.',
+            200
+        );
     }
 
     /**
@@ -79,11 +87,13 @@ class RehabilitationController extends Controller
      */
     public function destroy(Rehabilitation $rehabilitation)
     {
+        $this->authorize('delete', $rehabilitation);
+
         $this->rehabilitationService->delete($rehabilitation);
 
         return ApiResponse::success(
-           
-        ["message"=> 'Rehabilitation event deleted successfully.']
+
+            ["message" => 'Rehabilitation event deleted successfully.']
         );
     }
 }
